fix: remove MsgToL2 code and rework MsgToL1 hash function
