+++ starknet-core/src/types/mod.rs
@@ -54,8 +54,6 @@ pub mod requests;
 pub mod contract;
 pub use contract::ContractArtifact;
 
-pub mod messaging;
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum MaybePendingBlockWithTxHashes {
